Remove obsolete interface
