+++ score/launch_manager/src/daemon/src/watchdog/details/BUILD
@@ -13,14 +13,6 @@
 load("@rules_cc//cc:defs.bzl", "cc_library")
 
 cc_library(
-    name = "i_device_config_factory",
-    hdrs = ["IDeviceConfigFactory.hpp"],
-    include_prefix = "score/mw/launch_manager/watchdog",
-    strip_include_prefix = "/score/launch_manager/src/daemon/src/watchdog/details",
-    visibility = ["//score/launch_manager/src/daemon/src/watchdog/details:__subpackages__"],
-)
-
-cc_library(
     name = "watchdog",
     hdrs = ["Watchdog.hpp"],
     include_prefix = "score/mw/launch_manager/watchdog/details",
@@ -52,7 +44,6 @@ cc_library(
         "//score/launch_manager/src/daemon/src/common:log",
         "//score/launch_manager/src/daemon/src/configuration:config",
         "//score/launch_manager/src/daemon/src/watchdog:i_watchdog_if",
-        "//score/launch_manager/src/daemon/src/watchdog/details:i_device_config_factory",
         "@score_baselibs//score/language/futurecpp",
     ],
 )
